feat: refactor CLI for integration into AsyncAPI CLI (#1952)
